Adds set!
`set!(w::Widget, fld::Symbol, val)`

Set the value of a widget property and update all displayed instances of the
widget. If `val` is a `Signal`, then updates to that signal will be reflected in
widget instances/views.

If `fld` is `:value`, `val` is also `push!`ed to `signal(w)`